added fully coded Postgres repository, rest controller and much more. I had fun today :). it's just playing with the code from the course - in this project I don't need nice commits on GitHub

+++ src/main/java/ovh/major/songify/song/domain/model/SongEntity.java
@@ -9,10 +9,10 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-public class SongEntity {
+public class SongEntity{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     @Column(nullable = false)
     private String name;
     @Column(nullable = false)
